test: modernize to PHPUnit 11-13, reorganize test structure
- Upgrade phpunit.xml.dist to PHPUnit 11+ schema (11.5)
- Reorganize tests: test/Unit/Legacy/, test/Db/
- Separate DB persistence tests (test/Db/) from business logic tests
- Update BackendTestCase to use PHPUnit\Framework\TestCase
- Add getConfig() helper for legacy test compatibility
- Fix PHP 8.4 dynamic property deprecation in FileTest
- Add PHPUnit 11.5 to require-dev
- Create modern test/bootstrap.php
- Update namespaces: Horde\Token\Test\Unit\Legacy, Horde\Token\Test\Db
- All 20 unit tests passing (FileTest)
- Database tests properly skipped when adapters unavailable

Architectural improvement: Separating DB persistence concerns from
domain logic tests to enable cleaner PSR-4 conversion with proper
separation of concerns.
